Lint lib/
The carriers file contains duplicate keys. They have been resolved:
- exact duplicate of ideacellular deleted
- exact duplicate of onlinebeep deleted
- one duplicate of surewestcommunications (%s@freesurf.ch) renamed to
  sunrisecommunications as www.freesurf.ch redirects to sunrise.ch

Unused variables were defined in the library. They have been resolved:
- reference to the `underscore` module deleted
- reference to the `exec` function of the `child_process` module deleted

The carrier argument in text.send was not documented in the library
code. That line of documentation has been added.

The carrier argument in text.send was not being checked as strictly
equal to null. The triple-equals operator has now been applied to
the comparison.

The requires had bad line breakings before commas. The commas have
all been moved to ends of lines instead.

On an unrelated not to lint, the console.log call in the text library
has been switched out with an output call to obey the debug setting.